address reviewer comments, fix offload, change tags, move csv_iterator to common
