+++ pose_lookup/data_loading.hpp
@@ -1,5 +1,5 @@
 #include "common/error_util.hpp"
-#include "csv_iterator.hpp"
+#include "common/csv_iterator.hpp"
 
 #include <eigen3/Eigen/Dense>
 #include <fstream>
